buncha patches by darktjm sdl fix not applied yet
git-svn-id: https://svn.code.sf.net/p/vbam/code/trunk@1017 a31d4220-a93d-0410-bf67-fe4944624d44

+++ src/sdl/debugger.cpp
@@ -302,26 +302,26 @@ static void debuggerPrintBaseType(Type *t, u32 value, u32 location,
     case DW_ATE_signed:
       switch(debuggerRadix) {
       case 0:
-        printf("%lld", value);
+        printf("%lld", (long long)value);
         break;
       case 1:
-        printf("0x%llx", value);
+        printf("0x%llx", (long long)value);
         break;
       case 2:
-        printf("0%llo", value);
+        printf("0%llo", (long long)value);
         break;
       }
       break;
     case DW_ATE_unsigned:
       switch(debuggerRadix) {
       case 0:
-        printf("%llu", value);
+        printf("%llu", (unsigned long long)value);
         break;
       case 1:
-        printf("0x%llx", value);
+        printf("0x%llx", (unsigned long long)value);
         break;
       case 2:
-        printf("0%llo", value);
+        printf("0%llo", (unsigned long long)value);
         break;
       }
       break;
